Interests screen: replaces UI tests with screenshot tests
